Replace Double instances with BigDecimal instances for monetary handling

+++ app/src/main/java/com/jimtang/myshare/calc/CumulativeShareResults.java
@@ -1,11 +1,13 @@
 package com.jimtang.myshare.calc;
 
+import java.math.BigDecimal;
+
 /**
  * Created by tangz on 9/9/2015.
  */
 class CumulativeShareResults extends ShareResults {
 
-    public CumulativeShareResults(Double subtotal, Double tax, Double tip) {
+    public CumulativeShareResults(BigDecimal subtotal, BigDecimal tax, BigDecimal tip) {
         super("TOTAL", subtotal, tax, tip);
     }
 }
